lab2/requirements.txt
markdown
flask
google-generativeai